requirements.txt
requests
python-dotenv
customtkinter
pyperclip
